// frontend/src/app/core/api/model/sell-request.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SellRequest { 
    transactionId?: string;
    platform?: SellRequest.PlatformEnum;
    transactionDate?: string;
    assetId?: string;
    quantity?: number;
    price?: number;
    currency?: string;
    fee?: number;
}
export namespace SellRequest {
    export const PlatformEnum = {
        Trading212: 'TRADING212',
        Etoro: 'ETORO',
        Ibkr: 'IBKR',
        Revolut: 'REVOLUT'
    } as const;
    export type PlatformEnum = typeof PlatformEnum[keyof typeof PlatformEnum];
}