// frontend/src/app/core/api/model/dividend-request.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DividendRequest { 
    transactionId?: string;
    platform?: DividendRequest.PlatformEnum;
    transactionDate?: string;
    assetId?: string;
    grossAmount?: number;
    taxAmount?: number;
    taxPercentage?: number;
    currency?: string;
    netAmount?: number;
}
export namespace DividendRequest {
    export const PlatformEnum = {
        Trading212: 'TRADING212',
        Etoro: 'ETORO',
        Ibkr: 'IBKR',
        Revolut: 'REVOLUT'
    } as const;
    export type PlatformEnum = typeof PlatformEnum[keyof typeof PlatformEnum];
}